Move SignedTreeHead in 'uk.gov.register.proof.ct' package
